+++ .gitignore
@@ -15,4 +15,4 @@ bin/
 
 /text-ui-test/ACTUAL.txt
 text-ui-test/EXPECTED-UNIX.TXT
-data.txt
+*.txt
