--- src/main/java/duke/Task.java
@@ -1,32 +1,0 @@
-package duke;
-
-public class Task {
-    protected String description;
-    protected boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getStatusIcon() {
-        return (this.isDone ? "X" : " "); // mark done task with X
-    }
-
-    public void markAsDone() {
-        this.isDone = true;
-    }
-
-    public void markAsUndone() {
-        this.isDone = false;
-    }
-
-    public char getType() {
-        return '0';
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[%s] %s", getStatusIcon(), this.description);
-    }
-}
