+++ src/main/java/duke/exception/DukeInvalidCommandException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeInvalidCommandException extends DukeException {
     public DukeInvalidCommandException() {
