+++ src/main/java/duke/task/Task.java
@@ -1,0 +1,40 @@
+package duke.task;
+
+public class Task {
+    private String description;
+    private boolean isDone;
+
+    public Task(String description) {
+        description = description;
+        isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone() {
+        isDone = true;
+    }
+
+    public void markAsUndone() {
+        isDone = false;
+    }
+
+    public char getType() {
+        return '0';
+    }
+
+    public boolean getIsDone() {
+        return isDone;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s] %s", getStatusIcon(), this.description);
+    }
+}
